refactor: disambiguate historical modules, extract config serialization
Phase E of the cleanup plan.

Rename the three same-named "historical" modules so each states its layer:
- uk_data/adapters/historical.py     → adapters/historical_quarterly.py
                                        (low-level quarterly fetchers)
- companies_house_abm/abm/historical.py → abm/historical_simulation.py
                                        (the HistoricalSimulation runner)
- calibration/historical.py keeps its name; the package path already
  identifies it as the HistoricalAdapter orchestration layer.

Repoint every importer, mock patch-string, docstring, notebook, and doc
reference to the new module names.

CLI tidy: move the ModelConfig→YAML serialization out of cli.py into
abm/config.py as save_config()/config_to_dict(), next to load_config(),
and have the fetch-data command call it. Removes the bespoke
_write_calibrated_yaml/_to_dict helpers from the CLI.

make verify (55 warnings, unchanged), mkdocs build, and the offline test
suite (904 passed) all pass.
