+++ notebooks/historical_simulation.py
@@ -28,7 +28,7 @@ def _():
 def _(mo):
     import matplotlib.pyplot as plt
 
-    from companies_house_abm.abm.historical import HistoricalSimulation
+    from companies_house_abm.abm.historical_simulation import HistoricalSimulation
     from companies_house_abm.abm.scenarios import build_uk_2013_2024
 
     mo.md("## Setup")
